Se crea el fichero collectionForms con constantes de los formularios
