+++ src/pages/PlayerProfile/EditAttributes/EditAttributes.tsx
@@ -14,36 +14,13 @@ import {
   IonToolbar,
   useIonToast,
 } from "@ionic/react";
+import { checkboxList } from "constants/collectionsForms";
 import { EditTacticalAttributes } from "firebase/client";
 import { warningSharp } from "ionicons/icons";
 import { useState } from "react";
 import { useHistory } from "react-router";
 import styles from "./styles.module.css";
 
-const checkboxList = [
-  { val: "Fuerza", isChecked: false },
-  { val: "Entradas precisas", isChecked: false },
-  { val: "Capitán", isChecked: false },
-  { val: "Marcaje", isChecked: false },
-  { val: "Ritmo", isChecked: false },
-  { val: "Centros", isChecked: false },
-  { val: "Reacción", isChecked: false },
-  { val: "Destructor", isChecked: false },
-  { val: "Interceptor", isChecked: false },
-  { val: "Pases largos", isChecked: false },
-  { val: "Todocampista", isChecked: false },
-  { val: "Regateador", isChecked: false },
-  { val: "Control del balón", isChecked: false },
-  { val: "Tiro lejano", isChecked: false },
-  { val: "Contraataques", isChecked: false },
-  { val: "Ofensivo", isChecked: false },
-  { val: "Velocidad", isChecked: false },
-  { val: "Agilidad", isChecked: false },
-  { val: "Tiros", isChecked: false },
-  { val: "Finalización", isChecked: false },
-  { val: "Segundo Delantero", isChecked: false },
-  { val: "Referencia", isChecked: false },
-];
 const EditAttributes: React.FC = () => {
   const [present] = useIonToast();
   const [check, setCheck] = useState(true);
